prepare server for microservices architecture
