+++ app/lib/presentation/core/widgets/connection_state_listener_widget.dart
@@ -1,6 +1,6 @@
-import 'package:flutter/material.dart' hide ConnectionState;
-// ConnectionState hided because ConnectionBloc already has a class with the same name
+// ConnectionState hid because ConnectionBloc already has a class with the same name
 // renaming the class or using a namespace for connection_bloc are also possible options
+import 'package:flutter/material.dart' hide ConnectionState;
 import 'package:flutter_bloc/flutter_bloc.dart';
 
 import '../../../application/connection/connection_bloc.dart';
@@ -8,7 +8,7 @@ import '../../../l10n/l10n.dart';
 import 'simple_snackbar.dart';
 
 /// A top level widget that listens connection state
-/// Must be located under the MetarialApp context to be able to show snackbar messages
+/// Must be located under the MaterialApp context to be able to show snackbar messages
 class ConnectionStateListenerWidget extends StatelessWidget {
   /// ConnectionStateListenerWidget
   const ConnectionStateListenerWidget({required this.child});
@@ -21,8 +21,10 @@ class ConnectionStateListenerWidget extends StatelessWidget {
     return BlocListener<ConnectionBloc, ConnectionState>(
       listener: (context, state) {
         state.map(
-          connected: (_) => ScaffoldMessenger.of(context)
-              .showSnackBar(SimpleSnackBar(context.l10n.connected)),
+          connected: (_) {
+            ScaffoldMessenger.of(context)
+                .showSnackBar(SimpleSnackBar(context.l10n.connected));
+          },
           connectedOnInitial: (_) {},
           connectedWithNoNetworkAccess: (_) {},
           disconnected: (_) => ScaffoldMessenger.of(context)
