--- server/api/app.go
@@ -1,7 +1,0 @@
-package api
-
-import "github.com/wzslr321/artiver/entity"
-
-type application struct {
-	users *entity.UserCollection
-}
