--- server/entity/entity.go
@@ -1,2 +1,0 @@
-package entity
-
