--- server/infrastructure/repository/user_mongodb.go
@@ -1,1 +1,0 @@
-package repository
