--- server/infrastructure/repository/user_mongodb_test.go
@@ -1,1 +1,0 @@
-package repository_test
