--- server/pkg/password/password.go
@@ -1,1 +1,0 @@
-package password
