--- server/usecase/user/interface.go
@@ -1,1 +1,0 @@
-package user
