+++ server/users/Makefile
@@ -16,23 +16,16 @@ clean:
 dependencies:
 	go mod download
   
-build: dependencies build-cmd
+build: dependencies build-users
 
-build-cmd:
-	go build -tags $(LIBRARY_ENV) -o ./bin/cmd cmd/main.go
+build-users:
+	go build -o ./bin/users cmd/app/main.go
 
 linux-binaries:
 	CGO_ENABLED=0 GOOS=linux go build -a -installsuffix cgo -tags "$(LIBRARY_ENV) netgo" -installsuffix netgo -o $(BIN_DIR)/api api/main.go
 	CGO_ENABLED=0 GOOS=linux go build -a -installsuffix cgo -tags "$(LIBRARY_ENV) netgo" -installsuffix netgo -o $(BIN_DIR)/search cmd/main.go
 
 ci: dependencies test
-
-build-mocks:
-	@go get github.com/golang/mock/gomock
-	@go install github.com/golang/mock/mockgen
-	@~/go/bin/mockgen -source=usecase/book/interface.go -destination=usecase/book/mock/book.go -package=mock
-	@~/go/bin/mockgen -source=usecase/user/interface.go -destination=usecase/user/mock/user.go -package=mock
-	@~/go/bin/mockgen -source=usecase/loan/interface.go -destination=usecase/loan/mock/loan.go -package=mock
 
 test:
 	go test -tags testing ./...
