+++ server/users/cmd/app/main.go
@@ -3,9 +3,11 @@ package main
 import (
 	"context"
 	"fmt"
-	"github.com/rs/cors"
-	"github.com/wzslr321/artiver/api"
-	"github.com/wzslr321/artiver/settings"
+	cors "github.com/rs/cors/wrapper/gin"
+	"github.com/wzslr321/artiver/server/user/entity"
+	"github.com/wzslr321/artiver/server/user/settings"
+	"go.mongodb.org/mongo-driver/mongo"
+	"go.mongodb.org/mongo-driver/mongo/options"
 	"log"
 	"net/http"
 	"os"
@@ -14,18 +16,60 @@ import (
 	"time"
 )
 
-func init() {
-	settings.InitSettings()
-	api.InitMongo()
+type application struct {
+	users *entity.UserCollection
 }
 
-func main() {
+var app *application
 
+func init() {
+	settings.InitSettings()
+	initMongo()
+}
+
+func initMongo() {
+
+	oc := options.Client().ApplyURI(settings.MongodbSettings.Uri)
+
+	client, err := mongo.NewClient(oc)
+	if err != nil {
+		log.Fatalf("Error occured while initializing a new mongo client: %v", err)
+	}
+
+	ctx, cancel := context.WithTimeout(context.Background(), 15*time.Second)
+	defer cancel()
+
+	err = client.Connect(ctx)
+	if err != nil {
+		log.Fatalf("Errorr occurred while connecting to a client: %v", err)
+	}
+
+	defer func() {
+		if err = client.Disconnect(ctx); err != nil {
+			panic(err)
+		}
+	}()
+
+	log.Println("Successfully connected to the database!")
+
+
+	app = &application{
+		users: &entity.UserCollection{
+			C: client.Database("artiver").Collection("users"),
+		},
+	}
+}
+
+
+func main() {
 	address := fmt.Sprintf(":%s", settings.ServerSettings.Address)
-	router := cors.Default().Handler(api.App.InitRouter())
+	router := app.InitRouter()
 	readTimeout := settings.ServerSettings.ReadTimeout
 	writeTimeout := settings.ServerSettings.WriteTimeout
 	maxHeaderBytes := settings.ServerSettings.MaxHeaderBytes
+
+	router.Use(cors.Default())
+
 
 	server := &http.Server{
 		Addr:           address,
@@ -49,6 +93,7 @@ func main() {
 	}()
 
 	quit := make(chan os.Signal)
+
 
 	signal.Notify(make(chan os.Signal, 1), syscall.SIGINT, syscall.SIGTERM)
 	<-quit
