+++ server/users/cmd/app/router.go
@@ -1,4 +1,4 @@
-package api
+package main
 
 import (
 	"github.com/gin-gonic/gin"
