+++ server/users/go.mod
@@ -1,0 +1,13 @@
+module github.com/wzslr321/artiver/server/user
+
+go 1.16
+
+require (
+	github.com/gin-gonic/gin v1.5.0 // indirect
+	github.com/rs/cors v1.8.0 // indirect
+	github.com/stretchr/testify v1.7.0 // indirect
+	github.com/wzslr321/artiver v0.0.0-20210718173055-75401b6f2259 // indirect
+	go.mongodb.org/mongo-driver v1.7.0 // indirect
+	golang.org/x/crypto v0.0.0-20200302210943-78000ba7a073 // indirect
+	gopkg.in/ini.v1 v1.62.0 // indirect
+)
