+++ server/users/presenter/user.go
@@ -1,11 +1,13 @@
 package presenter
 
-import "github.com/wzslr321/artiver/entity"
+import (
+	"github.com/wzslr321/artiver/entity"
+)
 
 type User struct {
-	ID 		   entity.ID	`json:"id"`
-	Email 	   string 		`json:"email"`
-	Username   string		`json:"username"`
+	ID       entity.ID `json:"id"`
+	Email    string    `json:"email"`
+	Username string    `json:"username"`
 }
 
 type Register struct {
